Proof of concept: building app in Juno container
This experiment sets up a docker container based on
´elementary/docker:juno-unstable´.
It installs dependencies for ´io.elementary.terminal´ and
drops to a shell so that the programmer can interact with meson/ninja.
The script `enter_juno.sh` builds and executes the container.
Additionally, it connects the container to the host X server instance.
That allows to start the app inside the container.

## Issues

- Warnings about being unable to launch `dbus-launch`
- missing theme and icons
- _zoom_ and _changing theme_ actions have no effect

These seem to be __integration issues__ due to container isolation.
I don't know enough about this to solve those right now.